// src/interface/group-interface.ts
export interface FormGroupInterface {
    RpeAutAlta: string;
    Nivel: number;
    NombreG: string;
    ChkRegSolic: boolean | null;
    ChkPend: boolean | null;
    ChkBusquedaBas: boolean | null;
    ChkBusquedaAvz: boolean | null;
    ChkUsuASP: boolean | null;
    ChkUsuATodos: boolean | null;
    ChkUsuMSP: boolean | null;
    ChkUsuMTodos: boolean | null;
    ChkUsuBSP: boolean | null;
    ChkUsuBTodos: boolean | null;
    ChkAdmonBitacora: boolean | null;
    ChkAdmonGroups: boolean | null;
    ChkProbsSoluciones: boolean | null;
    ChkAdmonEstads: boolean | null;
    ChkMensajesError: boolean | null;
    ChkDocumentos: boolean | null;
    ChkTipoProbs: boolean | null;
    ChkModulos: boolean | null;
    ChkSuplencias: boolean | null;
    ChkEliminarSolic: boolean | null;
    ChkGradosUrgencia: boolean | null;
    ChkEncuestas: boolean | null;
    ChkSolucionesParticulares: boolean | null;
    ChkAgrpModulos: boolean | null;
    chkRep_Admon01: boolean | null;
    chkRep_Admon02: boolean | null;
    chkRep_Admon03: boolean | null;
    chkRep_Admon04: boolean | null;
    chkRep_2_Admon01: boolean | null;
    chkRep_2_Admon02: boolean | null;
    chkRep_2_Admon03: boolean | null;
    chkRep_2_Admon04: boolean | null;
    chkRep_2_Admon05: boolean | null;
    chkRep_2_Admon06: boolean | null;
    chkRep_3_Admon01: boolean | null;
    chkRep_3_Admon02: boolean | null;
    chkRep_3_Admon03: boolean | null;
    chkRep_3_Admon04: boolean | null;
    chkRep_3_Admon05: boolean | null;
    chkRep_3_Admon06: boolean | null;
    chkRep_3_Admon07: boolean | null;
    chkRep_2_Admon07: boolean | null;
    chkRep_Coord01: boolean | null;
    chkRep_Coord02: boolean | null;
    chkRep_Coord03: boolean | null;
    chkRep_Superv01: boolean | null;
    chkRep_Superv02: boolean | null;
    chkRep_Superv03: boolean | null;
    chkRep_Monit01: boolean | null;
    chkRep_Monit02: boolean | null;
    chkRep_2_Admon08: boolean | null;
    chkRep_3_Admon08: boolean | null;
    ChkBaseCon: boolean | null;
    ChkExpPS: boolean | null;
    ChkExpDoc: boolean | null;
    ChkBCMensajesError: boolean | null;
    ChkChatCte: boolean | null;
    ChkChatEjec: boolean | null;
    ChkManipulacion: boolean | null;
    ChkRec1: boolean | null;
    ChkRec3: boolean | null;
    config_tespera: boolean | null;
    config_intro: boolean | null;
    config_mensajesaviso: boolean | null;
    config_banner_gral: boolean | null;
    config_banner_CEDAC: boolean | null;
    config_banner_SII: boolean | null;
    config_tescalacion: boolean | null;
    config_tatencionsolic: boolean | null;
    config_vbitacora: boolean | null;
    config_contador: boolean | null;
    config_escautomatica: boolean | null;
    config_bienvenida: boolean | null;
    config_ligas_interes: boolean | null;
    ChkSugerenciasVer: boolean | null;
    ChkSugerenciasBorrar: boolean | null;
    ChkCuentaP: boolean | null;
    ChkForoAcceso: boolean | null;
    ChkForoAdmin: boolean | null;
    comentarios: string;
}